// src/lib/api/generated/models/user.ts
/**
 * Generated by orval v8.4.1 🍺
 * Do not edit manually.
 * Hubuum REST API
 * OpenAPI documentation for the Hubuum REST service.
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  created_at: string;
  /** @nullable */
  email?: string | null;
  id: number;
  password: string;
  updated_at: string;
  username: string;
}